Refactor finder packages (#473)
