+++ src/cmds/repo.rs
@@ -1,7 +1,7 @@
 use crate::filesystem;
+use crate::finder::structures::{Opts as FinderOpts, SuggestionType};
 use crate::finder::{Finder, FinderChoice};
 use crate::git;
-use crate::structures::finder::{Opts as FinderOpts, SuggestionType};
 use anyhow::Context;
 use anyhow::Error;
 use std::fs;
