+++ src/structures/mod.rs
@@ -1,4 +1,3 @@
 pub mod cheat;
 pub mod config;
-pub mod finder;
 pub mod item;
